requirements.txt
python==3.12.10
simpy==4.1.1
matplotlib==3.9.3
scipy==1.14.1
numpy==2.2.1
opencv-python==4.10.0.84
requests==2.32.5
python-dotenv==1.2.1
pandas==3.0.0